requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
elasticsearch==8.15.0
flask==3.0.3
nltk==3.9.1
python-dotenv==1.0.1
pytest==8.3.2
